// src/Optimism/utils/types.ts
export type DexType = 
  | "uniswapv3" 
  | "velodromeslipstream"
  | "solidlyv3"
  | "velodromefinancev2";